+++ backend/src/modules/internal-api/internal-api.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { InternalApiCoreModule } from './internal-api-core.module';
 import { InternalEmailModule } from './features/email/internal-email.module';
+import { InternalTimelineModule } from './features/timeline/internal-timeline.module';
 
 @Module({
-    imports: [InternalApiCoreModule, InternalEmailModule],
+    imports: [InternalApiCoreModule, InternalEmailModule, InternalTimelineModule],
 })
 export class InternalApiModule { }